split out username and password into its own component
